🚚 : built_in_type のディレクトリ構造変更
