+++ lib/chapter02/built_in_type/index.dart
@@ -1,0 +1,18 @@
+//　参照先のファイル内の変数は全てプライベート変数として扱う(_を先頭につける)
+import 'number.dart';
+import 'string.dart';
+import 'bool.dart';
+import 'list.dart';
+import 'set.dart';
+import 'map.dart';
+import 'record.dart';
+
+void main() {
+  printNumber();
+  printString();
+  printBool();
+  printList();
+  printSet();
+  printMap();
+  printRecord();
+}
