--- lib/chapter02/index.dart
@@ -1,18 +1,0 @@
-//　参照先のファイル内の変数は全てプライベート変数として扱う(_を先頭につける)
-import 'built_in_type/number.dart';
-import 'built_in_type/string.dart';
-import 'built_in_type/bool.dart';
-import 'built_in_type/list.dart';
-import 'built_in_type/set.dart';
-import 'built_in_type/map.dart';
-import 'built_in_type/record.dart';
-
-void main() {
-  printNumber();
-  printString();
-  printBool();
-  printList();
-  printSet();
-  printMap();
-  printRecord();
-}
